# Reduce memory usage, and speedup