These are the changes that took me (together with https://gitlab.cern.ch/cp3-cms/bamboo/merge_requests/80) from
```
INFO:bamboo.analysismodules:168 plots defined in 140.83s, max RSS: 6506.68MB
INFO:bamboo.analysismodules:Plots finished in 771.58s, max RSS: 12657.62MB (3108 histograms)
```
to
```
INFO:bamboo.analysismodules:168 plots defined in 83.29s, max RSS: 3812.42MB
INFO:bamboo.analysismodules:Plots finished in 532.87s, max RSS: 8056.09MB (3108 histograms)
```
The 3rd commit is the most interesting: by cutting on the MET after the b-jets, the cuts/variables that use the jets do not need to be calculated twice.